BIO-1506 - move dsmc logic into separate PDCClient class that can be mocked in local environments
