+++ archive_verify/pdc_client.py
@@ -1,0 +1,93 @@
+import re
+import logging
+import subprocess
+
+log = logging.getLogger(__name__)
+
+
+class BasePdcClient():
+    """
+    Base class representing a PDC client.
+    Staging and production environments should instantiate PdcClient.
+    Local and testing environments should instantiate MockPdcClient.
+    Default is PdcClient.
+    """
+
+
+class PdcClient(BasePdcClient):
+    @staticmethod
+    def download(archive, description, dest, dsmc_log_dir, whitelist):
+        """
+        Downloads the specified archive from PDC to a unique location.
+
+        :param archive: The path in PDC TSM to the archive that we want to download
+        :param description: The unique description used when previously uploaded the archive
+        :param dest: The unique path where we shall store the downloaded archive
+        :param dsmc_log_dir: Path to dir where we can save dsmc logs
+        :param whitelist: A list of dsmc warnings which we shall allow to be generated by dsmc but still count the download as successful
+        :returns True if no errors or only whitelisted warnings were encountered, False otherwise
+        """
+        log.debug("download_from_pdc started for {}".format(archive))
+        cmd = "export DSM_LOG={} && dsmc retr {}/ {}/ -subdir=yes -description='{}'".format(dsmc_log_dir, archive, dest,
+                                                                                            description)
+        p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+
+        dsmc_output, _ = p.communicate()
+        dsmc_exit_code = p.returncode
+
+        if dsmc_exit_code != 0:
+            return PdcClient._parse_dsmc_return_code(dsmc_exit_code, dsmc_output, whitelist)
+
+        log.debug("download_from_pdc completed successfully for {}".format(archive))
+        return True
+
+    @staticmethod
+    def _parse_dsmc_return_code(exit_code, output, whitelist):
+        """
+        Parses the dsmc output when we've encountered a non-zero exit code. For some certain exit codes,
+        warnings and errors we still want to return successfully.
+
+        :param exit_code: The exit code recieved from the failing dsmc process
+        :param output: The text output from the dsmc process
+        :param whitelist: A list of whitelisted warnings
+        :returns True if only whitelisted warnings was encountered in the output, otherwise False
+        """
+        log.debug("DSMC process returned an error!")
+
+        # DSMC sets return code to 8 when a warning was encountered.
+        if exit_code == 8:
+            log.debug("DSMC process actually returned a warning.")
+
+            output = output.splitlines()
+
+            # Search through the DSMC log and see if we only have
+            # whitelisted warnings. If that is the case, change the
+            # return code to 0 instead. Otherwise keep the error state.
+            warnings = []
+
+            for line in output:
+                matches = re.findall(r'ANS[0-9]+W', line)
+
+                for match in matches:
+                    warnings.append(match)
+
+            log.debug("Warnings found in DSMC output: {}".format(set(warnings)))
+
+            for warning in warnings:
+                if warning not in whitelist:
+                    log.error("A non-whitelisted DSMC warning was encountered. Reporting it as an error! ('{}')".format(
+                        warning))
+                    return False
+
+            log.debug("Only whitelisted DSMC warnings were encountered. Everything is OK.")
+            return True
+        else:
+            log.error("An uncaught DSMC error code was encountered!")
+            return False
+
+
+# TODO: Implement
+class MockPdcClient(BasePdcClient):
+     @staticmethod
+     def download(archive, description, dest, dsmc_log_dir, whitelist):
+         log.debug(f"Invoked MockPdcClient with parameters: {archive} {description} {dest} {dsmc_log_dir} {whitelist}")
