+++ archive_verify/workers.py
@@ -1,77 +1,13 @@
 import logging
-import re
 import rq
 import subprocess
 import os
 import datetime
 
+from archive_verify.pdc_client import PdcClient, MockPdcClient
+
 log = logging.getLogger(__name__)
 
-def _parse_dsmc_return_code(exit_code, output, whitelist):
-    """
-    Parses the dsmc output when we've encountered a non-zero exit code. For some certain exit codes, 
-    warnings and errors we still want to return successfully. 
-
-    :param exit_code: The exit code recieved from the failing dsmc process
-    :param output: The text output from the dsmc process
-    :param whitelist: A list of whitelisted warnings
-    :returns True if only whitelisted warnings was encountered in the output, otherwise False
-    """
-    log.debug("DSMC process returned an error!")
-
-    # DSMC sets return code to 8 when a warning was encountered.
-    if exit_code == 8:
-        log.debug("DSMC process actually returned a warning.")
-
-        output = output.splitlines()
-
-        # Search through the DSMC log and see if we only have
-        # whitelisted warnings. If that is the case, change the
-        # return code to 0 instead. Otherwise keep the error state.
-        warnings = []
-
-        for line in output:
-            matches = re.findall(r'ANS[0-9]+W', line)
-
-            for match in matches:
-                warnings.append(match)
-
-        log.debug("Warnings found in DSMC output: {}".format(set(warnings)))
-
-        for warning in warnings:
-            if warning not in whitelist:
-                log.error("A non-whitelisted DSMC warning was encountered. Reporting it as an error! ('{}')".format(warning))
-                return False
-
-        log.debug("Only whitelisted DSMC warnings were encountered. Everything is OK.")
-        return True
-    else:
-        log.error("An uncaught DSMC error code was encountered!")
-        return False
-
-def download_from_pdc(archive, description, dest, dsmc_log_dir, whitelist):
-    """
-    Downloads the specified archive from PDC to a unique location. 
-
-    :param archive: The path in PDC TSM to the archive that we want to download  
-    :param description: The unique description used when previsouly uploaded the archive
-    :param dest: The unique path where we shall store the downloaded archive
-    :param dsmc_log_dir: Path to dir where we can save dsmc logs 
-    :param whitelist: A list of dsmc warnings which we shall allow to be generated by dsmc but still count the download as successful
-    :returns True if no errors or only whitelisted warnings were encountered, False otherwise 
-    """
-    log.debug("download_from_pdc started for {}".format(archive))
-    cmd = "export DSM_LOG={} && dsmc retr {}/ {}/ -subdir=yes -description='{}'".format(dsmc_log_dir, archive, dest, description)
-    p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-
-    dsmc_output, _ = p.communicate()
-    dsmc_exit_code = p.returncode
-
-    if dsmc_exit_code != 0:
-        return _parse_dsmc_return_code(dsmc_exit_code, dsmc_output, whitelist)
-
-    log.debug("download_from_pdc completed successfully for {}".format(archive))
-    return True
 
 def compare_md5sum(archive_dir):
     """
@@ -93,6 +29,7 @@ def compare_md5sum(archive_dir):
     else: 
         return True
 
+
 def verify_archive(archive, archive_path, description, config):
     """
     Our main worker function. This will be put into the RQ/Redis queue when the /verify endpoint gets called. 
@@ -107,6 +44,8 @@ def verify_archive(archive, archive_path, description, config):
     dest_root = config["verify_root_dir"]
     dsmc_log_dir = config["dsmc_log_dir"]
     whitelist = config["whitelisted_warnings"]
+    pdc_client = MockPdcClient if config.get("pdc_client", "PdcClient") == "MockPdcClient" else PdcClient
+    log.debug(f"Using PDC Client of type: {pdc_client}")
 
     now_str = datetime.datetime.now().strftime('%Y-%m-%d_%H-%M')
     log.setLevel(logging.DEBUG)
@@ -121,7 +60,7 @@ def verify_archive(archive, archive_path, description, config):
     job_id = rq.get_current_job().id
     dest = "{}_{}".format(os.path.join(dest_root, archive), job_id)
 
-    download_ok = download_from_pdc(archive_path, description, dest, dsmc_log_dir, whitelist)
+    download_ok = pdc_client.download(archive_path, description, dest, dsmc_log_dir, whitelist)
 
     if not download_ok:
         log.debug("Download of {} failed.".format(archive))
